+++ libraries/fault_sdk/faultManagement.h
@@ -15,14 +15,14 @@ extern "C" {
 
 #include <fm_util.h>
 
-void faultManagement(FmState *fmState, float *rwSpeedHist, float *timeStampHisty, uint16_t rwDataLength, 
-		float *commandedTorque, double *fineDelTheta, double *coarseDelTheta, uint16_t sensorDataLength, float MOI);
+void faultManagement(FmState *fmState, double *rwSpeedHist, double *timeStampHisty, uint16_t rwDataLength, 
+		double *commandedTorque, double *fineDelTheta, double *coarseDelTheta, uint16_t sensorDataLength, double MOI);
 
 void manageNewFaultDetected(FmState *fmState);
 
-uint8_t faultCheckRW(FmState *fmState, float *frictionTorque, float *commandedTorque, uint16_t length);
+uint8_t faultCheckRW(FmState *fmState, double *frictionTorque, double *commandedTorque, uint16_t length);
 
-uint8_t checkThreshold(float *data_x, float *data_y, uint16_t length, float threshold);
+uint8_t checkThreshold(double *data_x, double *data_y, uint16_t length, double threshold);
 
 void manageFaultAlreadyDetected(FmState *fmState);
 
@@ -33,8 +33,8 @@ unsigned char faultCheck();
 // Performs numerical differentiation to determine angular acceleration 
 // by taking the difference of omega / difference of t. then multiplies by Moment of intertia (MOI)
 // to get the response torque on the reaction wheel
-void getResponseTorque(float *omega, float *t, float *responseTorque, float *commandedTorque, 
-	float *frictionTorque, uint16_t length, float MOI);
+void getResponseTorque(double *omega, double *t, double *responseTorque, double *commandedTorque, 
+	double *frictionTorque, uint16_t length, double MOI);
 
 unsigned char recovery(FmState *fmState);
